pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "ibm-watsonx-orchestrate"
dynamic = ["version"]

description = "IBM watsonx.orchestrate SDK"
authors = [
    { name = "IBM", email = "support@ibm.com" }
]
#readme = { file = "README.md", content-type = 'text/markdown'}
license = {text = "MIT License"}
license-files = ["LICENSE"]
requires-python = ">=3.11, <3.14"
classifiers = []
dependencies = [
    "certifi>=2024.8.30",
    "click>=8.0.0,<8.2.0",
    "docstring-parser>=0.16,<1.0",
    "httpx>=0.28.1,<1.0.0",
    "ibm-cloud-sdk-core>=3.22.0",
    "jsonref==1.1.0",
    "jsonschema>=4.23.0,<5.0.0",
    "langchain-community>=0.3.12,<1.0.0",
    "numpy>=1.26.0",
    "packaging>=24.2",
    "pydantic>=2.10.3,<3.0.0",
    "pyjwt>=2.10.1,<3.0.0",
    "python-dotenv>=1.0.0",
    "pyyaml>=6.0.2,<7.0.0",
    "requests>=2.32.0",
    "rich>=13.9.4,<14.0.0",
    "typer>=0.15.1,<1.0.0",
    "urllib3>=2.2.3",
    "munch>=4.0.0",
    "pytz>=2025.2",
    "redis>=6.0.0",
    "ibm_watsonx_orchestrate_evaluation_framework==1.0.2"
]

[project.optional-dependencies]
dev = [
    "pytest>=8.3.4,<9.0.0",
    "pytest-cov==6.0.0",
    "snapshottest==1.0.0a1",
    "pytest-mock==3.14.0",
    "pytest-asyncio==0.25.1",
    "coverage[toml]>=6.5",
    "black~=22.3.0",
    "pylint~=2.16.4",
]

[tool.hatch.envs.default]
dependencies = [
    "pkg[dev]"
]
detached = true

[tool.hatch.version]
path = "src/ibm_watsonx_orchestrate/__init__.py"
validate-bump=false


[tool.hatch.build.targets.wheel]
packages = ["src/ibm_watsonx_orchestrate"]

[tool.hatch.build]
include = [
    "src/*"
]

[project.scripts]
orchestrate = "ibm_watsonx_orchestrate.cli.main:app"

[tool.hatch.envs.default.scripts]
test = "pytest {args:tests}"
test-cov = "coverage run -m pytest {args:tests}"
cov-report = [
  "- coverage combine",
  "coverage report",
]
cov = [
  "test-cov",
  "cov-report",
]

[tool.coverage.run]
omit = [
  "*/__init__.py",
  "tests/*"
]

[tool.hatch.envs.test]
dependencies = [
    "pkg[dev]"
]

[[tool.hatch.envs.test.matrix]]
python = ["3.11", "3.12", "3.13"]

[tool.coverage.report]
show_missing=true
exclude_lines = [
  "no cov",
  "if __name__ == .__main__.:",
  "if TYPE_CHECKING:",
]
